agenticai-main-poc/elastic_confluence_rag/requirements.txt
langgraph
langgraph-supervisor
langchain-core
langchain-community
langchain-google-genai
google-generativeai
elasticsearch
python-dotenv
requests
